influxdb_to_mssql_statistics/requirements.txt
influxdb
pymssql
paho-mqtt
numpy
scipy
